Turned off "cross-products" for the "H" flag.
Because I just took out all "H"'s on verbs with ending flags
this is BASICALLY meaningless -- BUT one word (obainn) is,
by coincidence, a verb with 'E' affix flag and separately
a noun with an 'H' flag.   If cross-products remain
on, then illegal h+verb forms will still be included for this
one example!